data-collector/requirements.txt
yfinance
pandas
psycopg2-binary
python-dotenv
finance-datareader
selenium

